pyproject.toml
[tool.poetry]
name = "sistema-verificacao-inseges"
version = "2.0.0"
description = "Sistema de Verificação IN SEGES 65/2021 - Conferência automatizada de pesquisa de preços"
authors = ["Seu Nome <seu.email@exemplo.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.8"
streamlit = "^1.46.1"
flask = "^3.1.1"
flask-cors = "^6.0.0"
pandas = "^2.3.0"
pdfplumber = "^0.11.7"
PyMuPDF = "^1.26.1"
python-docx = "^1.2.0"
reportlab = "^4.4.2"
pillow = "^11.2.1"
google-generativeai = "^0.8.1"
python-dotenv = "^1.0.0"
bcrypt = "^4.1.2"
requests = "^2.32.4"
redis = {version = "^5.0.1", optional = true}
celery = {version = "^5.3.4", optional = true}

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
pytest-asyncio = "^0.21.1"
mypy = "^1.7.1"
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"
pre-commit = "^3.5.0"

[tool.poetry.extras]
cache = ["redis"]
async = ["celery"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py38']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --cov=src --cov-report=html --cov-report=term-missing"
testpaths = [
    "tests",
]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.coverage.run]
source = ["src"]
omit = ["*/tests/*", "*/venv/*", "*/__pycache__/*"]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false 